Moved unit tests to NN.cs
