+++ test/TorchSharpTest/NN.cs
@@ -11,6 +11,7 @@ using static TorchSharp.torch.nn.functional;
 using System.Drawing;
 using System.Collections.Generic;
 using System.Xml.Schema;
+using TorchSharp.Utils;
 
 #nullable enable
 
@@ -3020,6 +3021,27 @@ namespace TorchSharp
             }
         }
 
+
+        [Fact]
+        public void TestCustomComponentName()
+        {
+            var model = new TestCustomNameModel("TestCustomNameModel");
+
+            var sd = model.state_dict();
+            // Make sure that it's saved in the state_dict correctly, with and without custom attribute
+            Assert.True(sd.ContainsKey("custom_linear.weight"));
+            Assert.True(sd.ContainsKey("_linear2.weight"));
+
+            // The field names are also retrieved in the `_toEpilogue` function, so we want to make sure
+            // that everything works after calling a `.to` function. 
+            model = model.to(ScalarType.BFloat16);
+
+            sd = model.state_dict();
+            // Make sure that it's saved in the state_dict correctly, with and without custom attribute
+            Assert.True(sd.ContainsKey("custom_linear.weight"));
+            Assert.True(sd.ContainsKey("_linear2.weight"));
+        }
+
         private class TestModule3 : Module<Tensor, Tensor>
         {
             public TestModule3() : base(nameof(TestModule3)) { RegisterComponents(); }
@@ -3063,6 +3085,26 @@ namespace TorchSharp
             private Modules.Linear mod1 = Linear(10, 10);
             private Modules.Linear mod2 = Linear(10, 10);
             private Parameter p1 = new Parameter(torch.zeros(5, 5, 5), true);
+        }
+
+        private class TestCustomNameModel : nn.Module<Tensor, Tensor>
+        {
+            [ComponentName(Name = "custom_linear")]
+            private Linear _linear1;
+            private Linear _linear2;
+
+            public TestCustomNameModel(string name) : base(name)
+            {
+                _linear1 = Linear(5, 5, hasBias: false);
+                _linear2 = Linear(5, 5, hasBias: false);
+
+                RegisterComponents();
+            }
+
+            public override Tensor forward(Tensor input)
+            {
+                return _linear2.forward(_linear1.forward(input));
+            }
         }
         #endregion
